// user-service/src/models/user-model.ts
import mongoose, { Schema, type InferSchemaType } from 'mongoose';
import { USERNAME_REGEX } from '../utils/regex';

export const Roles = ['user', 'admin'] as const;
export type Role = (typeof Roles)[number];

export const SkillLevels = ['beginner', 'intermediate', 'advanced'] as const;
export type SkillLevel = (typeof SkillLevels)[number];

const userSchema = new Schema(
    {
        // Unique, immutable user id is automatically generated by MongoDB as _id

        // Used for login and management purposes, unique and case-insensitive
        username: {
            type: String,
            required: true,
            trim: true,
            lowercase: true,
            minlength: 3,
            maxlength: 30,
            unique: true,
            match: USERNAME_REGEX, // Disallow spaces
        },

        // Used for display purposes, not unique and case-sensitive
        displayName: { type: String, required: true, trim: true, maxlength: 50 },

        email: { type: String, required: true, trim: true, lowercase: true, unique: true },

        passwordHash: { type: String, required: true },

        role: { type: String, enum: Roles, default: 'user' },

        preferredLanguages: { type: [String], default: [] },

        skillLevel: { type: String, enum: SkillLevels, default: 'beginner' },

        refreshTokenHash: { type: String, default: null },

        refreshTokenIssuedAt: { type: Date, default: null },
    },
    { timestamps: true }, // Add createdAt and updatedAt fields automatically
);

userSchema.set('toJSON', {
    transform: (_doc, ret) => {
        const obj: any = { ...ret };
        obj.id = obj._id.toString(); // Expose _id as id
        delete obj.__v; // Hide version key
        delete obj._id; // Hide MongoDB's internal _id field
        delete obj.passwordHash; // Hide password hash
        delete obj.refreshTokenHash; // Hide refresh token hash
        delete obj.refreshTokenIssuedAt; // Hide refresh token issued at timestamp
        return obj;
    },
});

export type User = InferSchemaType<typeof userSchema>;
export const UserModel = mongoose.model('User', userSchema);